% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choR.R
\name{ChoR.as.cliques}
\alias{ChoR.as.cliques}
\title{Get the cliques.}
\usage{
ChoR.as.cliques(x)
}
\arguments{
\item{x}{A chordalysis object obtained by a call to ChoR.}
}
\value{
A list of cliques, a clique being a list of attributes'name, i.e. a list of lists of names.
}
\description{
Get the list of cliques associated to a chordalysis object.
}
