% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choR.R
\name{ChoR.as.formula}
\alias{ChoR.as.formula}
\title{Get the formula.}
\usage{
ChoR.as.formula(x)
}
\arguments{
\item{x}{A chordalysis object obtained by a call to ChoR.}
}
\value{
a formula representing the model
}
\description{
Extract the formula from a Chordalysis object.
}

