% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choR.R
\name{ChoR.as.graph}
\alias{ChoR.as.graph}
\title{Get the graph.}
\usage{
ChoR.as.graph(x)
}
\arguments{
\item{x}{A chordalysis object obtained by a call to ChoR.}
}
\value{
A graph
}
\description{
Get an undirected graph representing the cliques from a Chordalysis object.
}
\details{
The undirected graph use the graph package from Bioconductor.
}
