% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choR.R
\name{ChoR.loadData}
\alias{ChoR.loadData}
\title{[INTERNAL] Load the data from a dataframe (and with an optionnal cardinality vector)}
\usage{
ChoR.loadData(x, card = NULL)
}
\arguments{
\item{x}{A dataframe with categorical data; column names are the name of the attributes.}

\item{card}{A vectore containing the cardinality of the attributes (position wise).}
}
\value{
A list how two .jarray references (one for the dimension, one for the data) and the dataframe
}
\description{
Loads the data from x, which should be a dataframe (else, a conversion to a dataframe is attempted).
}
\details{
Loads the data from x, which should be a dataframe (else, a conversion to a dataframe is attempted).
The data must be categorical, each column being an attribute. The optionnal argument card should
be a vector representing the cardinality of each attribute (position wise).
If it is provided, its size must be equal to the number of attributes.
Else, its values will be computed from the data, and the cardinality for an attribute will be accurate only
if all its possible values appear at least once in the data.
}
