% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choR.R
\name{ChoR.processResult}
\alias{ChoR.processResult}
\title{[INTERNAL] Process the result of a java Chordalysis algorithm.}
\usage{
ChoR.processResult(x, modelStr)
}
\arguments{
\item{x}{The dataframe used to loadData; column names are the name of the attributes.}

\item{modelStr}{The result of a java Chordalysis algorithm}
}
\value{
A Chordalysis object. Use \code{ChoR.as.*} functions to access the result.
}
\description{
Convert the result in a 'chordalysis object'.
}
\details{
Process the result of a call to the java Chordalysis algorithm.
The result is a String of the forme "~0*1*2+...+3*4*5".
The numbers (+1 for indice correction) are replaced with the corresponding column name in x,
and the string is split in a list of cliques, a cliques being a list of name.
For example, "~ 0*1*2 + 3*4*5" gives the two cliques [[ [[0,1,2]], [[3,4,5]] ]]
}
