% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choR.R
\name{toString}
\alias{toString}
\title{[INTERNAL] Gives a string representation of the model.}
\usage{
toString(x)
}
\arguments{
\item{x}{A "Chordalysis" model, obtained by a call to a ChoR function.}
}
\value{
A String representation of the model.
}
\description{
Create a String representation of a model, compatible with the formula interface,
e.g. "~a*b*c+...+e*f*g".
}

